feat:⭐:Eh agregado una confirmación para poder eliminar una unidad o un conductor

+++ src/pages/drivers/Drivers.jsx
@@ -11,6 +11,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { useEffect, useMemo, useRef, useState } from "react";
 import { createDriver, deleteDriver, getDriverById, getDrivers, updateDriver } from "../../features/driverSlice";
 import * as Yup from "yup";
+import Swal from "sweetalert2";
 import "ag-grid-community/styles/ag-grid.css";
 import "ag-grid-community/styles/ag-theme-alpine.css";
 
@@ -93,7 +94,21 @@ const People = () => {
               <button
                 className="bg-red-500 hover:bg-red-400 text-white font-bold py-2 px-4 rounded-lg"
                 onClick={() => {
-                  dispatch(deleteDriver(value));
+                  Swal.fire({
+                    title: "¿Estás seguro?",
+                    text: "¡No podrás revertir esto!",
+                    icon: "warning",
+                    showCancelButton: true,
+                    confirmButtonColor: "#d33",
+                    cancelButtonColor: "#3085d6",
+                    confirmButtonText: "Sí, eliminar",
+                    cancelButtonText: "Cancelar",
+                  }).then((result) => {
+                    if (result.isConfirmed) {
+                      dispatch(deleteDriver(value));
+                    }
+                  }
+                  );
                 }}
               >
                 <TrashIcon className="h-6 w-6" />
